Refactor separation of main code and exports into two different files
